requirements.txt
discord.py
python-dotenv
tabulate
flask
firebase-admin